Moved the Config directory.
