+++ src/Utils/Config/Exception/File.php
@@ -12,7 +12,7 @@
  * @link https://github.com/jaxon-php/jaxon-core
  */
 
-namespace Jaxon\Config\Exception;
+namespace Jaxon\Utils\Config\Exception;
 
 class File extends \Exception
 {
