+++ src/Utils/Config/Exception/Yaml.php
@@ -13,7 +13,7 @@
  * @link https://github.com/jaxon-php/jaxon-core
  */
 
-namespace Jaxon\Config\Exception;
+namespace Jaxon\Utils\Config\Exception;
 
 class Yaml extends \Exception
 {
