+++ src/Utils/Config/Yaml.php
@@ -1,9 +1,9 @@
 <?php
 
 /**
- * Php.php - Jaxon config reader
+ * Yaml.php - Jaxon config reader
  *
- * Read the config data from a PHP config file.
+ * Read the config data from a YAML formatted config file.
  *
  * @package jaxon-core
  * @author Thierry Feuzeu <thierry.feuzeu@gmail.com>
@@ -12,12 +12,12 @@
  * @link https://github.com/jaxon-php/jaxon-core
  */
 
-namespace Jaxon\Config;
+namespace Jaxon\Utils\Config;
 
-class Php
+class Yaml
 {
     /**
-     * Read options from a PHP config file
+     * Read options from a YAML formatted config file
      *
      * @param string        $sConfigFile        The full path to the config file
      *
@@ -26,14 +26,18 @@ class Php
     public static function read($sConfigFile)
     {
         $sConfigFile = realpath($sConfigFile);
+        if(!extension_loaded('yaml'))
+        {
+            throw new \Jaxon\Utils\Config\Exception\Yaml(jaxon_trans('config.errors.yaml.install'));
+        }
         if(!is_readable($sConfigFile))
         {
-            throw new \Jaxon\Config\Exception\File(jaxon_trans('config.errors.file.access', ['path' => $sConfigFile]));
+            throw new \Jaxon\Utils\Config\Exception\File(jaxon_trans('config.errors.file.access', ['path' => $sConfigFile]));
         }
-        $aConfigOptions = include($sConfigFile);
+        $aConfigOptions = yaml_parse_file($sConfigFile);
         if(!is_array($aConfigOptions))
         {
-            throw new \Jaxon\Config\Exception\File(jaxon_trans('config.errors.file.content', ['path' => $sConfigFile]));
+            throw new \Jaxon\Utils\Config\Exception\File(jaxon_trans('config.errors.file.content', ['path' => $sConfigFile]));
         }
 
         return $aConfigOptions;
